6 read-write

+++ src/main/java/com/example/phone/repository/PhoneRepository.java
@@ -3,5 +3,5 @@ package com.example.phone.repository;
 import com.example.phone.model.Phone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhoneRepository extends JpaRepository<Phone, Long> {
+public interface PhoneRepository extends JpaRepository<Phone, Integer> {
 }
